pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "Spotter AI Dev"
authors = ["edmundrotimi"]
license = "MIT"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.11"
django = "3.2.23"
dotmap = "^1.3.30"
python-decouple = "^3.8"
django-split-settings = "^1.3.2"
sentry-sdk = {extras = ["django"], version = "^2.20.0"}
django-cleanup = "^9.0.0"
django-cors-headers = "4.5.0"
coverage = "^7.6.10"
django-autoslug = "^1.9.9"
pillow = "^11.1.0"
shortuuid = "^1.0.13"
django-csp = "^3.8"
django-simple-history = "3.6.0"
django-unfold = "^0.45.0"
djangorestframework = "3.15.1"
drf-yasg = "^1.21.8"
django-filter = "23.5"
django-defender = "^0.9.8"
django-auto-logout = "^0.5.1"
model-bakery = "1.17.0"
djangorestframework-simplejwt = "5.3.1"
django-import-export = "3.3.9"
requests = "^2.32.3"
jinja2 = "^3.1.5"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"
pre-commit = "^3.8.0"
django-debug-toolbar = "4.3.0"
werkzeug = "^3.0.4"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
